refactor: migrate FormItemSchemaToolbar
