+++ packages/core/client/src/modules/blocks/data-blocks/form/FormItemSchemaToolbar.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import { SchemaToolbar } from '../../schema-settings';
+import { SchemaToolbar } from '../../../../schema-settings';
 
 export const FormItemSchemaToolbar = (props) => {
   return <SchemaToolbar showBorder={false} showBackground {...props} />;
